# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CyberPulse AVA API - Enterprise Cybersecurity Companion Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateCurrentUser
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ─── CyberScore ─────────────────────────────────────────────────────────────
...
  /training/modules:
    get:
      operationId: getTrainingModules
      tags: [training]
      summary: List available security training modules
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [all, not_started, in_progress, completed]
            default: all
      responses:
        "200":
          description: Training modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingModule"
  /training/modules/{id}:
    get:
      operationId: getTrainingModule
      tags: [training]
      summary: Get training module details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Training module detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingModule"
...
  /training/modules/{id}/enroll:
    post:
      operationId: enrollTrainingModule
      tags: [training]
      summary: Enroll in or start a training module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Enrollment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingEnrollment"
  /training/modules/{id}/complete:
    post:
      operationId: completeTrainingModule
      tags: [training]
      summary: Mark a training module as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Completion record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingEnrollment"
  /training/leaderboard:
    get:
      operationId: getTrainingLeaderboard
      tags: [training]
      summary: Get top training achievers leaderboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

  # ─── AI Security Assistant ───────────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, displayName, email, role, department, avatarUrl, createdAt]
      properties:
        id:
          type: string
        displayName:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: string
        avatarUrl:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        theme:
          type: string
          enum: [light, dark, system]
        teamsWebhookUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
        theme:
          type: string
          enum: [light, dark, system]
        teamsWebhookUrl:
          type: ["string", "null"]

    # ─── CyberScore ────────────────────────────────────────────────────────────
...
    TrainingModule:
      type: object
      required: [id, title, category, durationMinutes, difficulty, status, completionRate]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        durationMinutes:
          type: integer
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        status:
          type: string
          enum: [not_started, in_progress, completed]
        completionRate:
          type: integer
          description: Global completion rate percentage
        userProgress:
          type: ["integer", "null"]
          description: User progress percentage 0-100
        scoreContribution:
          type: integer
          description: Points this module contributes to CyberScore
        tags:
          type: array
          items:
            type: string
        isMandatory:
          type: boolean
        dueDate:
          type: ["string", "null"]
...
    TrainingEnrollment:
      type: object
      required: [moduleId, userId, status, enrolledAt]
      properties:
        moduleId:
          type: string
        userId:
          type: string
        status:
          type: string
          enum: [enrolled, in_progress, completed]
        progress:
          type: integer
        enrolledAt:
          type: string
        completedAt:
          type: ["string", "null"]
    LeaderboardEntry:
      type: object
      required: [rank, userId, displayName, department, totalPoints, completedModules]
      properties:
        rank:
          type: integer
        userId:
          type: string
        displayName:
          type: string
        department:
          type: string
        avatarUrl:
          type: ["string", "null"]
        totalPoints:
          type: integer
        completedModules:
          type: integer

    # ─── AI Assistant ──────────────────────────────────────────────────────────
...