# packages/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CyberPulse AVA API - Enterprise Cybersecurity Companion Platform
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  displayName: string;
  department: string;
  /** @nullable */
  avatarUrl?: string | null;
  totalPoints: number;
  completedModules: number;
}