# packages/api-zod/src/generated/types/trainingEnrollment.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CyberPulse AVA API - Enterprise Cybersecurity Companion Platform
 * OpenAPI spec version: 0.1.0
 */
import type { TrainingEnrollmentStatus } from './trainingEnrollmentStatus';

export interface TrainingEnrollment {
  moduleId: string;
  userId: string;
  status: TrainingEnrollmentStatus;
  progress?: number;
  enrolledAt: string;
  /** @nullable */
  completedAt?: string | null;
}